Copied Speechly types and constants from browser-client to reduce size of the build
