+++ libraries/browser-ui/src/error-panel.svelte
@@ -1,7 +1,7 @@
 <svelte:options tag={null} immutable={true} />
 
 <script lang="ts">
-  import { ClientState, MessageType } from "./types";  // Re-exported from @speechl./fixTransitionclient. See types.ts for explanation.
+  import { ClientState, MessageType } from "./constants";  // Re-exported from @speechl./fixTransitionclient. See types.ts for explanation.
 
   const InvaldAppId = 'InvaldAppId'
   const HttpsRequired = 'HttpsRequired'
