+++ libraries/browser-ui/src/transcript-drawer.svelte
@@ -2,7 +2,7 @@
 
 <script lang="ts">
   import type { Segment } from '@speechly/browser-client';
-  import { ClientState, MessageType } from "./types";  // Re-exported from @speechly/browser-client. See types.ts for explanation.
+  import { ClientState, MessageType } from "./constants";  // Re-exported from @speechly/browser-client. See types.ts for explanation.
   import { cubicIn, cubicOut, linear } from 'svelte/easing';
   import { tweened } from 'svelte/motion';
   import "./big-transcript.ts";
